requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20

torch==2.8.0
torchvision==0.23.0
opencv-python==4.12.0.88
pillow==11.3.0
numpy==2.2.6
scikit-learn==1.7.2
scipy==1.16.2
pandas==2.3.2

ultralytics==8.3.199 